fix(middlewares): harden team and plantask middleware APIs
- Unexport internal types (MailboxSourceConfig, MailboxMessageSource,
  InboxMessage) to reduce public API surface
- Add ensureInit() guard to all exported Config methods preventing
  nil-state panics when called before NewRunner
- Fix data race on plantask reminder counter in team mode by using
  atomic operations
- Reject empty team name in CreateTeam with validation error
- Add regression tests: concurrent inbox writes, broadcast self-
  exclusion, corrupted JSON handling, name collision resolution,
  stale/corrupted highwatermark recovery, path-traversal ID rejection,
  concurrent reminder thread-safety

+++ adk/middlewares/plantask/plantask.go
@@ -223,7 +223,7 @@ type middleware struct {
 	// messages in state.Messages at the time onReminder was last invoked.
 	// Used to throttle subsequent reminders when onReminder is set, since the
 	// callback path does not inject a _task_reminder marker into messages.
-	lastCallbackReminderAssistantCount int
+	lastCallbackReminderAssistantCount int64
 
 	// Task assignment notification (set via WithTaskAssignedHook)
 	onTaskAssigned func(ctx context.Context, assignment TaskAssignment) error
